Added websocket server stuff to index.html.
Now the configuration pages talk to each other.

Added the feature such that when the master continues, it will prompt
all the slaves to continue (currently the "continue" url is google.com,
but this can be changed very easily.)

Also set up a system so it won't allow you to select a screen that has
already been selected.